requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.3.1
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
deprecation==2.1.0
fastapi==0.115.13
greenlet==3.2.3
grpcio==1.66.2
grpcio-health-checking==1.66.2
grpcio-tools==1.66.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
redis==5.0.1
idna==3.10
iniconfig==2.1.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.4.1
mypy_extensions==1.1.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.1
protobuf>=5.26.1,<6.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
ruff==0.12.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
structlog==25.4.0
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
validators==0.34.0
watchfiles==1.1.0
weaviate-client==4.15.2
websockets==15.0.1
xxhash==3.5.0
zstandard==0.23.0
google-cloud-aiplatform==1.97.0
google-auth==2.32.0
google-auth-oauthlib==1.2.0
google-cloud-firestore==2.21.0
google-cloud-storage==2.18.2
google-generativeai==0.8.3
google-cloud-speech==2.23.0
google-cloud-texttospeech==2.16.0

# Graph Database Options
# Option 1: Neo4j (if using Graphiti) - ~$275/month
# neo4j==5.25.0

# Option 2: Spanner Graph (GCP Native) - ~$100/month - RECOMMENDED
google-cloud-spanner==3.55.0

# LangChain dependencies for Spanner Graph (temporary)
langchain-google-vertexai==2.0.26
langchain-google-spanner==0.8.0

# HuggingFace for Gemma 2 9B
huggingface-hub==0.24.5
